(테스트 전) 필요 데이터 크롤러, DB 저장 모두 작성

+++ src/main/kotlin/kr/goldenmine/bus_improvement_crawler/requests/bus_stop/BusStopSectionResponseItem.kt
@@ -27,7 +27,7 @@ data class BusStopSectionResponseItem(
     val BSTOPSEQ: Int? = null,
     val DIRCD: Int? = null,
     val PATHSEQ: Int? = null,
-    val POSX: Double? = null,
-    val POSY: Double? = null,
+    val POSX: Float? = null,
+    val POSY: Float? = null,
     val SHORT_BSTOPID: Int? = null,
 )
